requirements.txt
# File transfer and loading utilities
hf_transfer
dotenv

# Core ML frameworks
torch
transformers
accelerate>=0.26.0
sentence-transformers
vllm>=0.6.2 # We need at least 0.6.2 to support LLaMA3.2-Vision. 

# Data processing and modeling
pandas
pydantic
datasets
Pillow
jinja2>=3.1.0
tokenizers

# API integrations
openai
cragmm-search-pipeline>=0.2.12
chromadb>=1.0.0

# Misc
loguru
rich
tqdm

